+++ test/FunctionMetadataGeneratorTests/FunctionMetadataGeneratorTests.cs
@@ -10,6 +10,7 @@ using System.IO;
 using System.Linq;
 using System.Reflection;
 using System.Threading.Tasks;
+using Azure.Messaging.EventHubs;
 using Azure.Storage.Blobs;
 using Microsoft.Azure.Functions.Tests;
 using Microsoft.Azure.Functions.Worker;
@@ -841,6 +842,59 @@ namespace Microsoft.Azure.Functions.SdkTests
             Assert.Throws<ArgumentOutOfRangeException>(() => new ExponentialBackoffRetryAttribute(5, "something_bad", "00:01:00"));
         }
 
+        [Fact]
+        public void EventHubs_SDKTypeBindings()
+        {
+            var generator = new FunctionMetadataGenerator();
+            var module = ModuleDefinition.ReadModule(_thisAssembly.Location);
+            var typeDef = TestUtility.GetTypeDefinition(typeof(SDKTypeBindings_EventHubs));
+            var functions = generator.GenerateFunctionMetadata(typeDef);
+            var extensions = generator.Extensions;
+
+            Assert.Equal(2, functions.Count());
+
+            AssertDictionary(extensions, new Dictionary<string, string>
+            {
+                { "Microsoft.Azure.WebJobs.Extensions.EventHubs", "5.4.0" },
+            });
+
+            var eventHubTriggerFunction = functions.Single(p => p.Name == nameof(SDKTypeBindings_EventHubs.EventHubTriggerFunction));
+
+            ValidateFunction(eventHubTriggerFunction, nameof(SDKTypeBindings_EventHubs.EventHubTriggerFunction), GetEntryPoint(nameof(SDKTypeBindings_EventHubs), nameof(SDKTypeBindings_EventHubs.EventHubTriggerFunction)),
+                ValidateEventHubTrigger);
+
+            var eventHubBatchTriggerFunction = functions.Single(p => p.Name == nameof(SDKTypeBindings_EventHubs.EventHubBatchTriggerFunction));
+
+            ValidateFunction(eventHubBatchTriggerFunction, nameof(SDKTypeBindings_EventHubs.EventHubBatchTriggerFunction), GetEntryPoint(nameof(SDKTypeBindings_EventHubs), nameof(SDKTypeBindings_EventHubs.EventHubBatchTriggerFunction)),
+                ValidateEventHubBatchTrigger);
+
+            void ValidateEventHubTrigger(ExpandoObject b)
+            {
+                AssertExpandoObject(b, new Dictionary<string, object>
+                {
+                    { "Name", "event" },
+                    { "Type", "eventHubTrigger" },
+                    { "Direction", "In" },
+                    { "eventHubName", "hub" },
+                    { "Cardinality", "One" },
+                    { "Properties", new Dictionary<string, object>( ) { { "SupportsDeferredBinding" , "True"} } }
+                });
+            }
+
+            void ValidateEventHubBatchTrigger(ExpandoObject b)
+            {
+                AssertExpandoObject(b, new Dictionary<string, object>
+                {
+                    { "Name", "events" },
+                    { "Type", "eventHubTrigger" },
+                    { "Direction", "In" },
+                    { "eventHubName", "hub" },
+                    { "Cardinality", "Many" },
+                    { "Properties", new Dictionary<string, object>( ) { { "SupportsDeferredBinding" , "True"} } }
+                });
+            }
+        }
+
         private class EventHubNotBatched
         {
             [Function("EventHubTrigger")]
@@ -1021,6 +1075,23 @@ namespace Microsoft.Azure.Functions.SdkTests
             public object BlobStringToBlobPocoArray(
                 [BlobTrigger("container2/%file%")] string blob,
                 [BlobInput("container2")] Poco[] blobinput)
+            {
+                throw new NotImplementedException();
+            }
+        }
+
+        private class SDKTypeBindings_EventHubs
+        {
+            [Function(nameof(EventHubTriggerFunction))]
+            public static void EventHubTriggerFunction(
+                [EventHubTrigger("hub", IsBatched = false)] EventData @event)
+            {
+                throw new NotImplementedException();
+            }
+
+            [Function(nameof(EventHubBatchTriggerFunction))]
+            public static void EventHubBatchTriggerFunction(
+                [EventHubTrigger("hub")] EventData[] events)
             {
                 throw new NotImplementedException();
             }